+++ 4TeethProfissional/app/src/main/java/br/com/ForTeethDentalCare/screens/MapsFragment.kt
@@ -1,19 +1,16 @@
-package br.com.ForTeethDentalCare
+package br.com.ForTeethDentalCare.screens
 
-import android.Manifest
-import android.content.pm.PackageManager
+import android.annotation.SuppressLint
 import android.location.Location
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
-import androidx.core.app.ActivityCompat
 import androidx.fragment.app.Fragment
+import br.com.ForTeethDentalCare.R
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.SupportMapFragment
 import com.google.android.gms.maps.model.LatLng
@@ -23,21 +20,15 @@ class MapsFragment : Fragment() {
 
     private lateinit var fusedLocationClient: FusedLocationProviderClient
 
+    @SuppressLint("MissingPermission")
     private val callback = OnMapReadyCallback { googleMap ->
-        if (ActivityCompat.checkSelfPermission(requireContext(), Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED ||
-            ActivityCompat.checkSelfPermission(requireContext(), Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-
-            fusedLocationClient.lastLocation
-                .addOnSuccessListener { location: Location? ->
-                    // Use the location object to get latitude, longitude, etc.
-                    val currentLocation = LatLng(location?.latitude!!, location.longitude)
-                    googleMap.addMarker(MarkerOptions().position(currentLocation).title("You are here"))
-                    googleMap.moveCamera(CameraUpdateFactory.newLatLng(currentLocation))
-                }
-        } else {
-            // Request necessary permissions or otherwise notify the user that the permissions are required
-            ActivityCompat.requestPermissions(requireActivity(), arrayOf(Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION), 0)
-        }
+        fusedLocationClient.lastLocation
+            .addOnSuccessListener { location: Location? ->
+                // Use the location object to get latitude, longitude, etc.
+                val currentLocation = LatLng(location?.latitude!!, location.longitude)
+                googleMap.addMarker(MarkerOptions().position(currentLocation).title("You are here"))
+                googleMap.moveCamera(CameraUpdateFactory.newLatLng(currentLocation))
+            }
     }
 
     override fun onCreateView(
