+++ 4TeethProfissional/app/src/main/java/br/com/ForTeethDentalCare/screens/NotificationsAreDisabledFragment.kt
@@ -1,0 +1,66 @@
+package br.com.ForTeethDentalCare.screens
+
+import android.Manifest
+import android.content.pm.PackageManager
+import android.os.Build
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
+import androidx.navigation.NavController
+import androidx.navigation.Navigation.findNavController
+import androidx.navigation.findNavController
+import br.com.ForTeethDentalCare.R
+import br.com.ForTeethDentalCare.databinding.FragmentNotificationsAreDisabledBinding
+
+class NotificationsAreDisabledFragment : Fragment() {
+
+    private var _binding: FragmentNotificationsAreDisabledBinding? = null
+    private val binding get() = _binding!!
+    private lateinit var navController: NavController
+
+    private val requestPermissionLauncher =
+        registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted ->
+            if (!isGranted) {
+                Toast.makeText(requireContext(), "Não foi possível habilitar as notificações", Toast.LENGTH_LONG)
+            }
+    }
+
+    private fun askNotificationPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            if (ContextCompat.checkSelfPermission(requireContext(), Manifest.permission.POST_NOTIFICATIONS) !=
+                PackageManager.PERMISSION_GRANTED
+            ) {
+                requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
+            }
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentNotificationsAreDisabledBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.btnAllowNotifications.setOnClickListener {
+            askNotificationPermission()
+        }
+    }
+
+    override fun onStart() {
+        super.onStart()
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
