+++ 4TeethProfissional/app/src/main/res/navigation/nav_graph_menu.xml
@@ -21,6 +21,9 @@
         <action
             android:id="@+id/menuFragment_to_serviceAddressesFragment"
             app:destination="@id/serviceAddressesFragment" />
+        <action
+            android:id="@+id/menuFragment_to_notificationsAreDisabledFragment"
+            app:destination="@id/notificationsAreDisabledFragment2" />
     </fragment>
     <fragment
         android:id="@+id/emergencyFragment"
@@ -63,11 +66,20 @@
     </fragment>
     <fragment
         android:id="@+id/mapsFragment"
-        android:name="br.com.ForTeethDentalCare.MapsFragment"
+        android:name="br.com.ForTeethDentalCare.screens.MapsFragment"
         android:label="fragment_maps"
         tools:layout="@layout/fragment_maps" >
         <action
             android:id="@+id/mapsFragment_to_userFragment"
             app:destination="@id/userFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/notificationsAreDisabledFragment2"
+        android:name="br.com.ForTeethDentalCare.screens.NotificationsAreDisabledFragment"
+        android:label="fragment_notifications_are_disabled"
+        tools:layout="@layout/fragment_notifications_are_disabled" >
+        <action
+            android:id="@+id/notificationsAreDisabledFragment_to_MenuFragment"
+            app:destination="@id/MenuFragment" />
     </fragment>
 </navigation>
